+++ tests/api/cli.py
@@ -71,7 +71,7 @@ class TestApi(unittest.TestCase):
                                        "--secret", self.secret])
 
         self.assertIsInstance(result.exception, DevoClientException)
-        self.assertEqual(result.exception.args[0]['status'], 401)
+        self.assertEqual(result.exception.args[0]['error']['code'], 12)
 
     def test_normal_query(self):
         runner = CliRunner()
@@ -85,7 +85,7 @@ class TestApi(unittest.TestCase):
 
         self.assertIsNone(result.exception)
         self.assertEqual(result.exit_code, 0)
-        self.assertIn('{"m":{"timestamp":{"type":"str","index":0}}}',
+        self.assertIn('{"m":{"timestamp":{"type":"str","index":0',
                       result.output)
 
     def test_with_config_file(self):
@@ -100,7 +100,7 @@ class TestApi(unittest.TestCase):
 
             self.assertIsNone(result.exception)
             self.assertEqual(result.exit_code, 0)
-            self.assertIn('{"m":{"timestamp":{"type":"str","index":0}}}',
+            self.assertIn('{"m":{"timestamp":{"type":"str","index":0',
                           result.output)